Fix random port allocation
In some circumstances jenkins allocates the same port multiple times during concurrent builds.
The jenkins tracks the allocated ports in a list but never checked the list when allocating a new port.
So, this fix fetches a new port from the underlying system when the current port is already in the allocated ports list.